recorder-backend/requirements.txt
backcall==0.1.0
boto3==1.12.18
botocore==1.15.18
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.2
docutils==0.15.2
idna==2.9
jedi==0.16.0
jmespath==0.9.5
Naked==0.1.31
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.4
ptyprocess==0.6.0
pycrypto==2.6.1
Pygments==2.6.1
python-dateutil==2.8.1
PyYAML==5.3
requests==2.23.0
s3transfer==0.3.3
shellescape==3.8.1
six==1.14.0
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.8
